AES-FF1 encrypt working with tests

+++ src/crypto/src/NIST.CVP.Crypto.Common/Symmetric/AES/IAesFfInternals.cs
@@ -1,4 +1,5 @@
 using NIST.CVP.Math;
+using System.Numerics;
 
 namespace NIST.CVP.Crypto.Common.Symmetric.AES
 {
@@ -19,7 +20,7 @@ namespace NIST.CVP.Crypto.Common.Symmetric.AES
         /// <param name="radix">The base.</param>
         /// <param name="x">The numeral string X.</param>
         /// <returns></returns>
-        short Num(short radix, NumeralString x);
+        BigInteger Num(int radix, NumeralString x);
 
         /// <summary>
         /// Algorithm 2 - Num
@@ -30,7 +31,7 @@ namespace NIST.CVP.Crypto.Common.Symmetric.AES
         /// </summary>
         /// <param name="x">The byte string X.</param>
         /// <returns></returns>
-        short Num(BitString x);
+        BigInteger Num(BitString x);
 
         /// <summary>
         /// Algorithm 3 - Str_m_radix
@@ -44,7 +45,7 @@ namespace NIST.CVP.Crypto.Common.Symmetric.AES
         /// <param name="m">the number of numerals.</param>
         /// <param name="x">The integer X, such that 0 &lt;= x &lt; radix^m  </param>
         /// <returns></returns>
-        NumeralString Str(short radix, short m, short x);
+        NumeralString Str(int radix, int m, BigInteger x);
 
         /// <summary>
         /// Given a numeral string, X, the numeral string that consists of the numerals
